FEAT-#2215: move implementation from experimental module
